Move slice cleanup controller
