+++ internal/controllers/resourceslice/slicecleanup_test.go
@@ -1,4 +1,4 @@
-package synthesis
+package resourceslice
 
 import (
 	"testing"
